test_requirements.txt
-e .

nose
rednose
mock
redis
python-memcached
